done fixing all errors
