+++ src/components/ModalWIthOrder/OrderDetails/OrderDetails.module.scss
@@ -1,0 +1,10 @@
+@use "../../../utils/mixins.scss" as m;
+
+.box {
+  @include m.flex();
+  flex-direction: column;
+  .glow {
+    text-shadow: 0px 0px 16px rgba(51, 51, 255, 0.25),
+      0px 0px 8px rgba(51, 51, 255, 0.25), 0px 4px 32px rgba(51, 51, 255, 0.5);
+  }
+}
