+++ src/index.jsx
@@ -7,7 +7,7 @@ import store from "./services/store";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  <React.StrictMode>
+  <React.StrictMode>``
     <Provider store={store}>
       <App />
     </Provider>
